--- python/ray/experimental/workflow/configs.py
@@ -1,5 +1,0 @@
-import pathlib
-
-
-def get_default_workflow_root_dir():
-    return pathlib.Path.cwd() / ".rayflow"
